+++ src/briefy/leica/models/__init__.py
@@ -8,6 +8,7 @@ from briefy.leica.models.asset import Asset
 from briefy.leica.models.asset import Image
 from briefy.leica.models.asset import ThreeSixtyImage
 from briefy.leica.models.asset import Video
+from briefy.leica.models.billing_info.customer import CustomerBillingInfo
 from briefy.leica.models.billing_info.professional import ProfessionalBillingInfo
 from briefy.leica.models.comment import Comment
 from briefy.leica.models.customer import Customer
@@ -52,6 +53,7 @@ ALL_MODELS = [
     Customer,
     CustomerUserProfile,
     CustomerBillingAddress,
+    CustomerBillingInfo,
     CustomerContact,
     Image,
     Facebook,