+++ HISTORY.rst
@@ -7,6 +7,8 @@ History
 
     * New model: ProfessionalBillingInfo (ericof).
     * New endpoint: /professionals/{professional_id}/billing (ericof).
+    * New model: CustomerBillingInfo (ericof).
+    * New endpoint: /customers/{customer_id}/billing (ericof).
     * On Order creation set order price based on project default value (ericof).